+++ server/internal/repository/pgadapter/pggen/channels.sql.go
@@ -277,7 +277,7 @@ type UpsertChannelParams struct {
 	OfflineImageUrl     *string `json:"offline_image_url"`
 	Description         *string `json:"description"`
 	BroadcasterType     *string `json:"broadcaster_type"`
-	ViewCount           int32   `json:"view_count"`
+	ViewCount           int64   `json:"view_count"`
 }
 
 func (q *Queries) UpsertChannel(ctx context.Context, arg UpsertChannelParams) (Channel, error) {
